requirements.txt
# Core dependencies
streamlit>=1.25.0
langchain>=0.0.267
langchain-core>=0.1.4
langchain-community>=0.0.12
langgraph>=0.0.21

# LLM & Generation
openai>=1.3.0
deepseek-sdk>=0.1.0  # Optional, but OpenAI SDK is used for DeepSeek

# Vector database
qdrant-client>=1.6.3

# Embedding models
sentence-transformers>=2.2.2
torch>=2.0.0

# Document processing
PyPDF2>=3.0.0
pdf2image>=1.16.3
docx2txt>=0.8
python-docx>=0.8.11

# Web Scraping
crawl4ai>=0.1.0
asyncio>=3.4.3

# Authentication & Security
bcrypt>=4.0.1
python-dotenv>=1.0.0

# Search Utilities
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0

# Utilities
huggingface-hub>=0.17.3
uuid>=1.30
logging>=0.5.1